ref(notifications): move join notification firing to notifications feature
